// types/product-detail-settings.ts
export interface ProductDetailWarrantySettings {
  warrantySectionEnabled: boolean
  warrantySectionTitle: string
  warrantyPrimaryTitle: string
  warrantyPrimaryDescription: string
  technicalSupportTitle: string
  technicalSupportDescription: string
  genuinePartsTitle: string
  genuinePartsDescription: string
  lifetimeSupportTitle: string
  lifetimeSupportDescription: string
}